Introducing the `use` key
The `use` key brings into one thing all the various ways of using a soft
path, so subsumes `restore`, `insert`, and `append` (and their
variants).  It allows for additional options to be specified which
control how a new path is inserted into an existing one (if there is an
existing one) and so should streamline making complicated paths out of
pieces without needing a lot of cloning and transforming keys.