Improvements to the gaps and joining code.
The gaps code now works without specifying the components list in which
case it inserts gaps between *all* components.

There is a new version of the code for joining components with a splice
path which tests to see if the gap is "upright" and if not inserts the
reflection of the joining path.