Fixed bug in checking previous/next filaments
When an intersection point is near the end of a filament, the
previous/next filament -- and possibly the one before that as well --
should be pre/ap-pended before the over path is rendered.  The tests for
existence of these was not being done correctly.

(This also contains the start of implementing the rectangle code for
soft paths.)