Default path is current, fix bug with fp comparisons
In the spath3 library, path actions that take a single argument now have
`current` as the default when appropriate.

In the knots library, I clearly thought that `\fp_compare` could do
multiple comparisons.  When the knots are embedded in some sort of
alignment environment then this caused it to fail dismally.  Using
`\bool_if` fixes the problem.