Disable arrow shortening
Shortening the path when it is baked can cause issues with arrows.
Putting an arrow in a path definition affects the path because the path
gets shortened so that the arrow ends where the path was meant to end.
So an arrow affects the path definition, but the arrow is not itself
part of the path so if an arrow is used when the path is defined and
again when the path is used, the path will be shortened twice which
might not be what is intended.  Therefore it is useful to have a way to
disable the shortening and place an arrow tip at the actual end of the
line.