Reset path bounding box when setting the current path
TikZ/PGF keeps track of two bounding boxes when a path is created: the
overall picture bounding box and a path bounding box. The command
`\pgf@protocol@sizes` keeps these up to date as the path is created.

When an spath3 path is used to replace the current path then these
bounding boxes should be updated accordingly. Simply invoking
`\pgf@protocol@sizes` is fine for the picture bounding box, but not for
the path bounding box since that retains the information of how the
original path was defined and that may have changed. So this commit adds
a reset for the path bounding box whenever spath3 sets the current path.

Note that there are some edge cases where, for example, it isn't
appropriate to set the picture bounding box. These can be dealt with by
judicious use of the `overlay` key. I'll add examples to the
documentation in due course.